api/requirements.txt
numpy
torch
torchvision
fastapi
uvicorn
Pillow
python-multipart